+++ src/components/ui/dropdown-menu/dropdown-menu.styles.ts
@@ -24,6 +24,6 @@ export const dropdownMenuStyles = {
   shortcut: "ml-auto text-xs tracking-widest text-gray-500",
   iconWrapper: "absolute left-2 flex h-3.5 w-3.5 items-center justify-center",
   chevronIcon: "ml-auto h-4 w-4 text-gray-500",
-  checkIcon: "h-4 w-4 text-teal-600",
-  circleIcon: "h-2 w-2 fill-teal-600"
+  checkIcon: "h-4 w-4 text-emerald-600",
+  circleIcon: "h-2 w-2 fill-emerald-600"
 } as const;